#6 ParallelException as nested class
