--- src/main/java/com/jcabi/aspects/aj/ParallelException.java
@@ -1,65 +1,0 @@
-/**
- * Copyright (c) 2012-2013, JCabi.com
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met: 1) Redistributions of source code must retain the above
- * copyright notice, this list of conditions and the following
- * disclaimer. 2) Redistributions in binary form must reproduce the above
- * copyright notice, this list of conditions and the following
- * disclaimer in the documentation and/or other materials provided
- * with the distribution. 3) Neither the name of the jcabi.com nor
- * the names of its contributors may be used to endorse or promote
- * products derived from this software without specific prior written
- * permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT
- * NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
- * FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL
- * THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
- * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
- * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
- * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
- * OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package com.jcabi.aspects.aj;
-
-/**
- * Exception that encapsulates all exceptions thrown from concurrent threads.
- *
- * @author Krzysztof Krason (Krzysztof.Krason@gmail.com)
- * @version $Id$
- * @since 0.10
- * @see Parallelizer
- */
-public final class ParallelException extends Exception {
-
-    /**
-     * Next parallel exception.
-     */
-    private final transient ParallelException next;
-
-    /**
-     * Constructor.
-     * @param cause Cause of the current exception.
-     * @param nxt Following exception.
-     */
-    public ParallelException(final Throwable cause,
-        final ParallelException nxt) {
-        super(cause);
-        this.next = nxt;
-    }
-
-    /**
-     * Get next parallel exception.
-     * @return Next exception.
-     */
-    public ParallelException getNext() {
-        return this.next;
-    }
-}
